Modify AL_AIR_ABSORPTION_FACTOR behavior
According to the specification, AL_AIR_ABSORPTION_FACTOR is applied (only) on the direct
path of the source. Tests made on Creative HW show this behavior for almost all effects,
with the exception of AL_EFFECT_EAXREVERB and AL_EFFECT_REVERB effects. On these effects,
the direct path is applied in the same way of the other effects and for the effect path,
the internal Air Absorption Gain HF indicated on the spec. is the parameter
AL_EAXREVERB_AIR_ABSORPTION_GAINHF or AL_REVERB_AIR_ABSORPTION_GAINHF and combines
multiplicatively (in dB) with AL_AIR_ABSORPTION_FACTOR.